Added four different kinds of plots
Prof. Kartalov requested plots that show the same material
in the same print direction before and after baking. This
commit adds new Type enums that describe the new plots
and adds the corresponding selections to execute the code.

I changed the output directory in config.py to one that
takes the Excel file from WSL2 Linux (instead of from a
path in WiNdows 10).