Big refactor - problem with design...
Tried to create Channel objects from the two different data sources, ImageJ CSV files and an Excel spreadsheet (AmScope data). The AmScope data are structured differently. I tried to solve this by specifying the data source in the constructor. This leads to problems in the code down stream. To resolve it I would need to turn good code into bad/hackish code. This needs re-thinking...

I think we should let the constructor figure out from the data passed to what the data source is. That's so far a better solution.